+++ GHEtool/Validation/artificial_neural_network.py
@@ -1,0 +1,30 @@
+"""
+The goal of this file is to validate the artificial neural network, based on the work of Blanke et al. ([#BlankeEtAl]_)
+
+References
+----------
+.. [#BlankeEtAl] Blanke T., Pfeiffer F., Göttsche J., Döring B. (2024) Artificial neural networks use
+for the design of geothermal probe fields. In Proceedings of BauSim Conference 2024:  10th Conference of
+IBPSA-Germany and Austria. Vienna (Austria), 23-26 September 2024. https://doi.org/10.26868/29761662.2024.12
+"""
+from GHEtool import *
+
+import time
+
+
+def validate_L2():
+    pass
+
+
+def validate_L3():
+    pass
+
+
+def validate_L4():
+    pass
+
+
+if __name__ == "__main__":  # pragma: no cover
+    validate_L2()
+    validate_L3()
+    validate_L4()