+++ GHEtool/Validation/sizing_method_comparison.py
@@ -60,7 +60,7 @@ def sizing_method_comparison():
 
     # create the borefield object
     borefield = Borefield(load=load)
-    borefield.set_ground_parameters(data)
+    borefield.ground_data = data
     borefield.set_borefield(borefield_gt)
     borefield.Rb = 0.2